# services/beeja-projects/src/main/resources/static/openApi.yaml
openapi: 3.0.1
info:
  title: Beeja Projects Management API
  version: v1.0
  description: |-
    `Product of tech.at.core`
    API documentation for the Beeja Projects Management Service. It allows users to manage
    projects, clients, and contracts via APIs secured by JWT authentication.
    
    ## Workflow Notes:
    1. First create Clients
    2. Then create Projects (requires clientId)
    3. Finally create Contracts (requires projectId)

    ## Authentication & Authorization
    Beeja uses JWT bearer token authentication.
    [Login to Beeja](https://app.beeja.io/login) to obtain your access token.

  contact:
    name: Beeja API Support
    email: api-support@beeja.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8000/projects
    description: Local development server
  - url: /projects/v1
    description: Relative path (works behind proxies)
security:
  - bearerAuth: []
...
paths:
...
  /v1/invoices:
    post:
      tags: [ Invoices ]
      summary: Create a new invoice for a contract
      operationId: createInvoice
      description: |
        Generates an invoice for a given contract. Requires permission: `CREATE_INVOICE`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceRequest'
      responses:
        201:
          description: Invoice created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        500:
          $ref: '#/components/responses/InternalServerError'
  /v1/invoices/{invoiceId}:
    get:
        tags: [ Invoices ]
        summary: Get invoice by ID
        operationId: getInvoiceById
        description: |
          Retrieves a single invoice by its unique ID. Requires permission: `GET_INVOICE`.
        parameters:
          - name: invoiceId
            in: path
            required: true
            schema:
              type: string
        responses:
          200:
            description: Invoice found
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Invoice'
          401:
            $ref: '#/components/responses/Unauthorized'
          403:
            $ref: '#/components/responses/Forbidden'
          404:
            $ref: '#/components/responses/NotFound'
          500:
            $ref: '#/components/responses/InternalServerError'
  /v1/invoices/contract/{contractId}:
    get:
        tags: [ Invoices ]
        summary: Get all invoices for a contract
        operationId: getInvoicesByContractId
        description: |
          Retrieves all invoices associated with a specified contract. Requires permission: `GET_INVOICE`.
        parameters:
          - name: contractId
            in: path
            required: true
            schema:
              type: string
        responses:
          200:
            description: List of invoices for the contract
            content:
              application/json:
                schema:
                  type: array
                  items:
                    $ref: '#/components/schemas/Invoice'
          401:
            $ref: '#/components/responses/Unauthorized'
          403:
            $ref: '#/components/responses/Forbidden'
          500:
            $ref: '#/components/responses/InternalServerError'
  /v1/invoices/{invoiceId}/mark-paid:
    put:
        tags: [ Invoices ]
        summary: Mark invoice as paid
        operationId: markInvoiceAsPaid
        description: |
          Updates an invoice's status to 'paid'. Requires permission: `UPDATE_STATUS_INVOICE`.
        parameters:
          - name: invoiceId
            in: path
            required: true
            schema:
              type: string
        responses:
          200:
            description: Invoice marked as paid successfully
            content:
              application/json:
                schema:
                  $ref: '#/components/schemas/Invoice'
          401:
            $ref: '#/components/responses/Unauthorized'
          403:
            $ref: '#/components/responses/Forbidden'
          404:
            $ref: '#/components/responses/NotFound'
          500:
            $ref: '#/components/responses/InternalServerError'
components:
  schemas:
...
    Invoice:
      type: object
      properties:
        id:
          type: string
          description: MongoDB document ID
        invoiceId:
          type: string
          description: Business identifier for the invoice
        contractId:
          type: string
        projectId:
          type: string
        clientId:
          type: string
        organizationId:
          type: string
        amount:
          type: number
          format: double
        currency:
          type: string
          example: "USD"
        billingDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [ PENDING, PAID, CANCELLED, OVERDUE ]
        notes:
          type: array
          items:
            type: string
        invoiceFileId:
          type: string
        paymentDetails:
          $ref: '#/components/schemas/PaymentDetails'
        remittanceRef:
          type: string
        taxId:
          type: string
        amountInWords:
          type: string
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        vat:
          type: integer
        daysLeftForPayment:
          type: string
        invoicePeriod:
          $ref: '#/components/schemas/InvoicePeriod'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    InvoicePeriod:
        type: object
        properties:
          startDate:
            type: string
            format: date-time
          endDate:
            type: string
            format: date-time
    PaymentDetails:
      type: object
      description: Payment details for the invoice
      properties:
        accountName:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        ifscNumber:
          type: string
    Task:
      type: object
      description: Task included in the invoice
      properties:
        taskName:
          type: string
        description:
          type: string
        price:
          type: number
          format: double
    Error:
      type: object
      properties:
        failureReason:
          type: string
          enum: [ CONFLICT, BAD_REQUEST, NOT_FOUND ]
        type:
          type: string
          example: "validation_error"
        code:
          type: string
          example: "BAD_REQUEST"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object
          example:
            field: "name"
            message: "must not be blank"
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          example:
            failureReason: UNAUTHORIZED
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          example:
            failureReason: FORBIDDEN
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          example:
            failureReason: NOT_FOUND
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Server error
      content:
        application/json:
          example:
            failureReason: INTERNAL_SERVER_ERROR
          schema:
            $ref: '#/components/schemas/Error'
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT